Fix percentage issue

+++ app/src/main/java/android/com/ericswpark/awipe/MainActivity.kt
@@ -165,7 +165,7 @@ class MainActivity : AppCompatActivity() {
         val progressBar = findViewById<ProgressBar>(R.id.main_activity_wipe_progress_bar)
         val progressText = findViewById<TextView>(R.id.main_activity_wipe_progress_text)
 
-        val percentage = currentBytes * 100 / totalBytes
+        val percentage = currentBytes.toDouble() * 100 / totalBytes
         progressBar.progress = percentage.toInt()
 
         progressText.text = String.format("%d/%d, %f%%", currentBytes, totalBytes, currentBytes.toDouble() / totalBytes)
